Automatically generate and save IDs
To ensure a consistent experience, especially if re-generating the module, IDs should be consistent across runs,
so store them in a file using a hash as the key (try to keep possibly-copyrighted content to a minimum) and load
that on startup. If we're running in dev, save it at the end.

Also add hashes of the files to test, as the files should now be identical every run.