Add handlebar templates for generating HTML
It's *significnatly* easier to read and edit than doing the HTML in Java. Also made a few other tweaks to make it easier to use handlebars.

Figured out how to solve the Record/GraalVM/Reflection problem - it needs all methods to be accessed via reflection, so added some test utilities to make sure that happens for classes that cause the build to fail.