Add remaining Handlebars templates and convert most classes to records
Found a workaround for GraalVM issue by using reflection to call all methods on a record. Very weird, but it works. Converted all data classes to records, because there's way, way less boilerplate.

Convert \r\n to \n in all Handlebars output, in the hope that it was this or encoding issues that were causing the Windows build to fail.