Add in data class generator
It reads the template information from a Foundry system and generates the various data classes it provides. Not 100% perfect, as it needs manual tweaking to handle things that should be enums, but good enough to save a ton of work.

Also provides a builder and sets defaults, so it's much easier to work with, especially multiple different items with the data type are needed (e.g., Fetish & Talen)